feat(orchestrate-core): move the playbook type model into the core (#108)
Second slice of the drive-core extraction (noetl/ai-meta#108): the playbook type
model (`Playbook`, `Step`, `ToolSpec`, `ToolKind`, `CursorClaim`, `Command`, …)
moves from `src/playbook/types.rs` into `noetl-orchestrate-core` — it's the input
to `evaluate`, so the core (native + wasm32) needs it, and `commands`/`evaluator`/
`state` all depend on it.

The model was pure serde with one coupling: `KeychainDef.residency:
crate::secrets::residency::Residency`.  The `Residency` enum is itself part of the
playbook model (a keychain's region policy) and pure data, so it moves into the
core too (with its one `as_label` helper, now `pub`); the server's
`secrets::residency` re-exports the enum and keeps the region-check *logic* (which
needs `AppError`/metrics/`server_region`).  `crate::playbook::types` is re-exported
from the core via `pub use noetl_orchestrate_core::playbook as types`, so all 7
call sites + the parser are unchanged.

Core deps: `serde` (derive) + `serde_yaml` (dev, for the moved type tests).
The model's ~19 unit tests moved with it.  632 server lib + 40 core tests green;
core still compiles to wasm32-unknown-unknown (no WASI); clippy clean.  Pure
refactor — no runtime change.

Next slice: `commands` (now both its deps — the renderer + the type model — live
in the core).

Refs noetl/ai-meta#108
